pyproject.toml
# Copyright 2019, by the California Institute of Technology.
# ALL RIGHTS RESERVED. United States Government Sponsorship acknowledged.
# Any commercial use must be negotiated with the Office of Technology
# Transfer at the California Institute of Technology.
#
# This software may be subject to U.S. export control laws. By accepting
# this software, the user agrees to comply with all applicable U.S. export
# laws and regulations. User has the responsibility to obtain export
# licenses, or other export authority as may be required before exporting
# such information to foreign countries or providing access to foreign
# persons.

[tool.poetry]
name = "l2ss-py"
version = "3.2.0a1"
description = "L2 Subsetter Service"
authors = ["podaac-tva <podaac-tva@jpl.nasa.gov>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/podaac/l2ss-py"
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules"
]
exclude = ['contrib', 'docs', 'tests']
packages = [
    { include = "podaac" },
]

[tool.poetry.dependencies]
python = "^3.12"
netCDF4 = "^1.5"
xarray = {version = "<=2025.1.0", extras = ["parallel"]}
geopandas ="^1.0.1"
Shapely = "^2.0.6"
harmony-service-lib = { version = "^2.0.0", optional = true }
pystac = { version = "^1.10.1", optional = true }
julian = "^0.14"
importlib-metadata = "^8.2.0"
h5py = "^3.6.0"
cf-xarray = "*"
numpy = "^2.2.4"
poetry = "^2.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
flake8 = "^7.0.0"
pytest-cov = "^6.0.0"
pylint = "^3.0.2"
sphinx = "^8.1.3"
pytest-benchmark = "^5.1.0"
moto = "^5.0.2"
jsonschema = "^4.20.0"
m2r2 = "^0.3.2"
sphinx-rtd-theme = "^3.0.2"
pytest-rerunfailures = "^15.0"
pytest-xdist = "^3.6.1"
myst-parser = "^4.0.1"

[tool.pytest.ini_options]
junit_family = "xunit2"

[tool.poetry.extras]
harmony = ["harmony-service-lib", "pystac"]

[tool.poetry.scripts]
l2ss_harmony = 'podaac.subsetter.subset_harmony:main'
l2ss-py = 'podaac.subsetter.run_subsetter:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
